requirements.txt
celery==5.3.4
Django==4.2.20
djangorestframework==3.15.0
gunicorn==21.2.0
ipython==8.11.0
model-bakery==1.20.0
model-bakery==1.20.0
openpyxl==3.1.5
pandas==2.2.0
pre-commit==4.0.0
psycopg2-binary==2.9.7
pytest==7.2.1
pytest-django==4.8.0
redis==4.5.4
virtualenv==20.16.3
